+++ Sources/Gzip/Data+Gzip.swift
@@ -5,7 +5,7 @@
 /*
  The MIT License (MIT)
  
- © 2014-2017 1024jp <wolfrosch.com>
+ © 2014-2019 1024jp <wolfrosch.com>
  
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
@@ -34,9 +34,7 @@ import Foundation
     import zlib
 #endif
 
-/**
- Compression level whose rawValue is based on the zlib's constants.
- */
+/// Compression level whose rawValue is based on the zlib's constants.
 public struct CompressionLevel: RawRepresentable {
     
     /// Compression level in the range of `0` (no compression) to `9` (maximum compression).
@@ -63,53 +61,40 @@ public struct CompressionLevel: RawRepresentable {
 }
 
 
-/**
- Errors on gzipping/gunzipping based on the zlib error codes.
- */
+/// Errors on gzipping/gunzipping based on the zlib error codes.
 public struct GzipError: Swift.Error {
     // cf. http://www.zlib.net/manual.html
     
-    public enum Kind {
-        /**
-         The stream structure was inconsistent.
-         
-         - underlying zlib error: `Z_STREAM_ERROR` (-2)
-         */
+    public enum Kind: Equatable {
+        /// The stream structure was inconsistent.
+        ///
+        /// - underlying zlib error: `Z_STREAM_ERROR` (-2)
         case stream
         
-        /**
-         The input data was corrupted (input stream not conforming to the zlib format or incorrect check value).
-         
-         - underlying zlib error: `Z_DATA_ERROR` (-3)
-         */
+        /// The input data was corrupted
+        /// (input stream not conforming to the zlib format or incorrect check value).
+        ///
+        /// - underlying zlib error: `Z_DATA_ERROR` (-3)
         case data
         
-        /**
-         There was not enough memory.
-         
-         - underlying zlib error: `Z_MEM_ERROR` (-4)
-         */
+        /// There was not enough memory.
+        ///
+        /// - underlying zlib error: `Z_MEM_ERROR` (-4)
         case memory
         
-        /**
-         No progress is possible or there was not enough room in the output buffer.
-         
-         - underlying zlib error: `Z_BUF_ERROR` (-5)
-         */
+        /// No progress is possible or there was not enough room in the output buffer.
+        ///
+        /// - underlying zlib error: `Z_BUF_ERROR` (-5)
         case buffer
         
-        /**
-         The zlib library version is incompatible with the version assumed by the caller.
-         
-         - underlying zlib error: `Z_VERSION_ERROR` (-6)
-         */
+        /// The zlib library version is incompatible with the version assumed by the caller.
+        ///
+        /// - underlying zlib error: `Z_VERSION_ERROR` (-6)
         case version
         
-        /**
-         An unknown error occurred.
-         
-         - parameter code: return error by zlib
-         */
+        /// An unknown error occurred.
+        ///
+        /// - parameter code: return error by zlib
         case unknown(code: Int)
     }
     
@@ -158,25 +143,19 @@ public struct GzipError: Swift.Error {
 
 extension Data {
     
-    /**
-     Whether the data is compressed in gzip format.
-     */
+    /// Whether the receiver is compressed in gzip format.
     public var isGzipped: Bool {
         
         return self.starts(with: [0x1f, 0x8b])  // check magic number
     }
     
     
-    /**
-     Create a new `Data` object by compressing the receiver using zlib.
-     Throws an error if compression failed.
-     
-     - parameters:
-     - level: Compression level.
-     
-     - throws: `GzipError`
-     - returns: Gzip-compressed `Data` object.
-     */
+    /// Create a new `Data` object by compressing the receiver using zlib.
+    /// Throws an error if compression failed.
+    ///
+    /// - Parameter level: Compression level.
+    /// - Returns: Gzip-compressed `Data` object.
+    /// - Throws: `GzipError`
     public func gzipped(level: CompressionLevel = .defaultCompression) throws -> Data {
         
         guard !self.isEmpty else {
@@ -219,13 +198,11 @@ extension Data {
     }
     
     
-    /**
-     Create a new `Data` object by decompressing the receiver using zlib.
-     Throws an error if decompression failed.
-     
-     - throws: `GzipError`
-     - returns: Gzip-decompressed `Data` object.
-     */
+    /// Create a new `Data` object by decompressing the receiver using zlib.
+    /// Throws an error if decompression failed.
+    ///
+    /// - Returns: Gzip-decompressed `Data` object.
+    /// - Throws: `GzipError`
     public func gunzipped() throws -> Data {
         
         guard !self.isEmpty else {